+++ src/com/yahoo/ycsb/generator/ScrambledZipfianGenerator.java
@@ -132,6 +132,6 @@ public class ScrambledZipfianGenerator extends IntegerGenerator
 	 */
 	@Override
 	public double mean() {
-		return ((double)(((long)_min) +(long)_max))/2.0;
+		return ((_min) +_max)/2.0;
 	}
 }
