+++ src/com/yahoo/ycsb/workloads/ConstantOccupancyWorkload.java
@@ -77,7 +77,7 @@ public class ConstantOccupancyWorkload extends CoreWorkload {
 		double fieldsize = g.mean();
 		int fieldcount = Integer.parseInt(p.getProperty(FIELD_COUNT_PROPERTY, FIELD_COUNT_PROPERTY_DEFAULT));
 
-		object_count = (long)(occupancy * ((double)disksize / (fieldsize * (double)fieldcount)));
+		object_count = (long)(occupancy * (disksize / (fieldsize * fieldcount)));
                 if(object_count == 0) {
                     throw new IllegalStateException("Object count was zero.  Perhaps disksize is too low?");
                 }
